+++ ontolib-cli/src/main/java/com/github/phenomics/ontolib/cli/MergeScoresCommand.java
@@ -1,8 +1,7 @@
 package com.github.phenomics.ontolib.cli;
 
-import com.github.phenomics.ontolib.ontology.scoredist.ScoreDistribution;
-import com.github.phenomics.ontolib.ontology.scoredist.ScoreDistributions;
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,8 +9,19 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.SortedMap;
 import java.util.TreeMap;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.github.phenomics.ontolib.base.OntoLibException;
+import com.github.phenomics.ontolib.base.OntoLibRuntimeException;
+import com.github.phenomics.ontolib.io.scoredist.H2ScoreDistributionWriter;
+import com.github.phenomics.ontolib.io.scoredist.ScoreDistributionReader;
+import com.github.phenomics.ontolib.io.scoredist.ScoreDistributionWriter;
+import com.github.phenomics.ontolib.io.scoredist.TextFileScoreDistributionReader;
+import com.github.phenomics.ontolib.io.scoredist.TextFileScoreDistributionWriter;
+import com.github.phenomics.ontolib.ontology.scoredist.ScoreDistribution;
+import com.github.phenomics.ontolib.ontology.scoredist.ScoreDistributions;
 
 /**
  * Implementation of {@code merge-scores} command.
@@ -53,7 +63,9 @@ public class MergeScoresCommand {
     LOGGER.info("Loading distributions...");
 
     for (String inputPath : options.getInputFiles()) {
-      try (ScoreDistributionReader reader = new ScoreDistributionReader(new File(inputPath))) {
+      LOGGER.info("Loading {}", inputPath);
+      try (final ScoreDistributionReader reader =
+          new TextFileScoreDistributionReader(new File(inputPath))) {
         Map<Integer, ScoreDistribution> distributions = reader.readAll();
         for (Entry<Integer, ScoreDistribution> e : distributions.entrySet()) {
           if (!loadedDists.containsKey(e.getKey())) {
@@ -61,14 +73,13 @@ public class MergeScoresCommand {
           }
           loadedDists.get(e.getKey()).add(e.getValue());
         }
-      } catch (IOException e) {
-        throw new RuntimeException("Problem reading input file: " + inputPath, e);
+      } catch (OntoLibException | IOException e) {
+        throw new OntoLibRuntimeException("Problem reading input file: " + inputPath, e);
       }
     }
 
     LOGGER.info("Done loading distributions.");
   }
-
 
   private void mergeDistributions() {
     LOGGER.info("Merging distributions...");
@@ -80,20 +91,35 @@ public class MergeScoresCommand {
     LOGGER.info("Done merging distributions.");
   }
 
-
   private void writeResult() {
-    LOGGER.info("Writing result file...");
-
-    try (ScoreDistributionWriter writer =
-        new ScoreDistributionWriter(new File(options.getOutputFile()))) {
+    LOGGER.info("Writing result...");
+    try (ScoreDistributionWriter writer = buildWriter()) {
       for (Entry<Integer, ScoreDistribution> e : mergedDists.entrySet()) {
         writer.write(e.getKey(), e.getValue(), 0);
       }
-    } catch (IOException e) {
+    } catch (IOException | OntoLibException e) {
       throw new RuntimeException("Problem writing to output file: " + options.getOutputFile(), e);
     }
+    LOGGER.info("Done writing result.");
+  }
 
-    LOGGER.info("Done writing result file.");
+  /**
+   * @return {@link ScoreDistributionWriter}, depending on configured path and type.
+   * @throws OntoLibException in the case of problems creating the writer.
+   */
+  private ScoreDistributionWriter buildWriter() throws OntoLibException {
+    if (options.isWriteToH2()) {
+      LOGGER.info("Creating H2 database connection for writing score distribution...");
+      final String pathDbAbs = new File(options.getOutputFile()).getAbsolutePath();
+      return new H2ScoreDistributionWriter(pathDbAbs, options.getH2TableName(), true);
+    } else {
+      LOGGER.info("Opening text file for writing score distribution...");
+      try {
+        return new TextFileScoreDistributionWriter(new File(options.getOutputFile()));
+      } catch (FileNotFoundException e) {
+        throw new OntoLibException("Could not find file " + options.getOutputFile(), e);
+      }
+    }
   }
 
   private void printHeader() {