+++ ontolib-cli/src/main/java/com/github/phenomics/ontolib/cli/package-info.java
@@ -2,7 +2,8 @@
  * Command line interface to Ontolib.
  *
  * <p>
- * Currently, this only gives access to score distribution precomputation.
+ * Currently, this only gives access to score distribution precomputation and merging of score
+ * files.
  * </p>
  *
  * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>